Fix a bunch of pylint issues
